cloud/gpu-worker/requirements.txt
# REM GPU Worker Dependencies

# AWS SDK
boto3>=1.34.0

# Whisper transcription
faster-whisper>=0.10.0

# OpenAI for embeddings and summarization
openai>=1.0.0

# Audio processing
numpy>=1.24.0
ffmpeg-python>=0.2.0

# Environment variables
python-dotenv>=1.0.0

# Logging
colorlog>=6.8.0

# Speaker diarization
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0

# For speaker embedding extraction and comparison
scipy>=1.10.0

